pyproject.toml
[project]
name = "interview-chatbot"
version = "0.1.0"
description = "Weather & Climate AI Chatbot using Pydantic AI and Open-Meteo APIs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-ai-slim[openrouter,web]>=1.56.0",
    "pydantic-ai>=1.56.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
